+++ packages/smarthr-ui/src/components/Dialog/MessageDialog/stories/MessageDialog.stories.tsx
@@ -1,29 +1,30 @@
+import { useRef, useState } from 'react'
 import { action } from 'storybook/actions'
 
-import type { Meta, StoryObj } from '@storybook/react'
-import { useRef, useState } from 'react'
 import { Button } from '../../../Button'
 import { MessageDialog } from '../MessageDialog'
 import { MessageDialogContent } from '../MessageDialogContent'
+
+import type { Meta, StoryObj } from '@storybook/react-webpack5'
 
 export default {
   title: 'Components/Dialog/MessageDialog',
   component: MessageDialog,
   subcomponents: { MessageDialogContent },
-  render: ({ onClickClose, ...args }) => {
+  render: ({ onClickClose, ...rest }) => {
     const [open, setOpen] = useState(false)
     const handleClose = onClickClose ?? (() => setOpen(false))
 
     return (
       <>
         <Button onClick={() => setOpen(true)}>ダイアログを開く</Button>
-        <MessageDialog {...args} onClickClose={handleClose} isOpen={open} />
+        <MessageDialog {...rest} onClickClose={handleClose} isOpen={open} />
       </>
     )
   },
   argTypes: {
-    titleTag: {
-      name: 'titleTag（非推奨）',
+    unrecommendedTitleTag: {
+      name: 'unrecommendedTitleTag（非推奨）',
     },
   },
   args: {
@@ -51,10 +52,10 @@ export const Subtitle: StoryObj<typeof MessageDialog> = {
   },
 }
 
-export const TitleTag: StoryObj<typeof MessageDialog> = {
-  name: 'titleTag（非推奨）',
+export const UnrecommendedTitleTag: StoryObj<typeof MessageDialog> = {
+  name: 'unrecommendedTitleTag（非推奨）',
   args: {
-    titleTag: 'h3',
+    unrecommendedTitleTag: 'h3',
   },
 }