+++ packages/smarthr-ui/src/components/Dropdown/SortDropdown/stories/SortDropdown.stories.tsx
@@ -3,7 +3,7 @@ import { action } from 'storybook/actions'
 import { Cluster } from '../../../Layout'
 import { SortDropdown } from '../SortDropdown'
 
-import type { Meta, StoryObj } from '@storybook/react'
+import type { Meta, StoryObj } from '@storybook/react-webpack5'
 
 export default {
   title: 'Components/Dropdown/SortDropdown',