+++ packages/smarthr-ui/src/components/Heading/PageHeading/PageHeading.tsx
@@ -4,12 +4,13 @@ import { type PropsWithChildren, memo, useEffect, useId, useMemo } from 'react'
 import innerText from 'react-innertext'
 import { tv } from 'tailwind-variants'
 
+import { IS_NEXT_JS } from '../../../libs/nextjs'
 import { STYLE_TYPE_MAP, Text, type TextProps } from '../../Text'
-import { VisuallyHiddenText, visuallyHiddenTextClassNameGenerator } from '../../VisuallyHiddenText'
+import { VisuallyHiddenText, visuallyHiddenTextClassName } from '../../VisuallyHiddenText'
 
 import type { ElementProps } from '../Heading'
 
-export type Props = PropsWithChildren<{
+export type AbstractProps = PropsWithChildren<{
   /**
    * テキストのサイズ
    *
@@ -18,13 +19,18 @@ export type Props = PropsWithChildren<{
   size?: Extract<TextProps['size'], 'XXL' | 'XL' | 'L'>
   /** 視覚的に非表示にするフラグ */
   visuallyHidden?: boolean
-  /** title要素の自動生成フラグ */
+  /**
+   * title要素の自動生成フラグ
+   *
+   * Next.js 環境ではこの値にかかわらずtitleは自動生成されません。metadataなどの方法を利用してください。
+   */
   autoPageTitle?: boolean
   /** title要素のprefix */
   pageTitle?: string
   /** title要素のsuffix */
   pageTitleSuffix?: string
 }>
+type Props = AbstractProps & Omit<ElementProps, keyof AbstractProps>
 
 const classNameGenerator = tv({
   base: 'smarthr-ui-Heading smarthr-ui-PageHeading',
@@ -38,9 +44,7 @@ const classNameGenerator = tv({
   },
 })
 
-const PSEUDO_TITLE_CLASS_NAME = visuallyHiddenTextClassNameGenerator()
-
-export const PageHeading = memo<Props & ElementProps>(
+export const PageHeading = memo<Props>(
   ({
     size,
     className,
@@ -49,7 +53,7 @@ export const PageHeading = memo<Props & ElementProps>(
     pageTitleSuffix = 'SmartHR（スマートHR）',
     pageTitle,
     children,
-    ...props
+    ...rest
   }) => {
     const actualClassName = useMemo(
       () => classNameGenerator({ visuallyHidden, className }),
@@ -66,26 +70,31 @@ export const PageHeading = memo<Props & ElementProps>(
     }, [size])
 
     const pseudoTitleId = useId()
-    const titleText = useMemo(
-      () => (autoPageTitle ? `${pageTitle || innerText(children)}｜${pageTitleSuffix}` : ''),
+
+    const autoTitleText = useMemo(
+      () =>
+        autoPageTitle && !IS_NEXT_JS
+          ? `${pageTitle || innerText(children)}｜${pageTitleSuffix}`
+          : '',
       [children, pageTitle, pageTitleSuffix, autoPageTitle],
     )
 
     useEffect(() => {
-      if (titleText) {
+      if (autoTitleText) {
+        document.title = autoTitleText
+
         // HINT: SPAで遷移する場合などの対策としてbody直下にaria-liveを仕込む
         // head内はスクリーンリーダーの変更検知のチェック対象外のため、title要素にaria-liveは設定しない
         const pseudoTitle: HTMLDivElement = (document.getElementById(pseudoTitleId) ||
           document.createElement('div')) as HTMLDivElement
 
         pseudoTitle.setAttribute('id', pseudoTitleId)
-        pseudoTitle.setAttribute('class', PSEUDO_TITLE_CLASS_NAME)
+        pseudoTitle.setAttribute('class', visuallyHiddenTextClassName)
         pseudoTitle.setAttribute('aria-live', 'polite')
         document.body.prepend(pseudoTitle)
 
-        document.title = titleText
         requestAnimationFrame(() => {
-          pseudoTitle.innerText = titleText
+          pseudoTitle.textContent = autoTitleText
         })
 
         return () => {
@@ -94,12 +103,12 @@ export const PageHeading = memo<Props & ElementProps>(
       }
 
       return undefined
-    }, [titleText, pseudoTitleId])
+    }, [autoTitleText, pseudoTitleId])
 
     const Component = visuallyHidden ? VisuallyHiddenText : Text
 
     return (
-      <Component {...props} {...actualTypography} as="h1" className={actualClassName}>
+      <Component {...rest} {...actualTypography} as="h1" className={actualClassName}>
         {children}
       </Component>
     )