+++ packages/smarthr-ui/src/components/InformationPanel/InformationPanel.tsx
@@ -25,9 +25,9 @@ type AbstractProps = PropsWithChildren<{
   /** パネルのタイトル */
   title: ReactNode
   /**
-   * @deprecated titleTagは非推奨です
+   * 可能な限り利用せず、SectioningContent(Article, Aside, Nav, Section)を使ってInformationPanel全体を囲むことで、InformationPanelのheadingのレベルを調整する方法を検討してください
    */
-  titleTag?: HeadingTagTypes
+  unrecommendedTitleTag?: HeadingTagTypes
   /** `true` のとき、開閉ボタンを表示する */
   toggleable?: boolean
   /** 開閉ボタン押下時に発火するコールバック関数 */
@@ -113,7 +113,7 @@ export const classNameGenerator = tv({
 
 export const InformationPanel: FC<Props> = ({
   title,
-  titleTag,
+  unrecommendedTitleTag,
   type = 'info',
   toggleable,
   active: activeProps = true,
@@ -122,7 +122,7 @@ export const InformationPanel: FC<Props> = ({
   children,
   onClickTrigger,
   decorators,
-  ...props
+  ...rest
 }) => {
   const [active, setActive] = useState(activeProps)
   const id = useId()
@@ -168,10 +168,14 @@ export const InformationPanel: FC<Props> = ({
   const classNames = classNamesMapper[active ? 'active' : 'inactive']
 
   return (
-    <Base {...props} overflow="hidden" as="section" className={classNames.wrapper}>
+    <Base {...rest} overflow="hidden" as="section" className={classNames.wrapper}>
       <Sidebar align="baseline" right className={classNames.header}>
-        {/* eslint-disable-next-line smarthr/a11y-heading-in-sectioning-content */}
-        <MemoizedHeading tag={titleTag} id={titleId} className={classNames.heading} type={type}>
+        <MemoizedHeading
+          unrecommendedTag={unrecommendedTitleTag}
+          id={titleId}
+          className={classNames.heading}
+          type={type}
+        >
           {title}
         </MemoizedHeading>
         {toggleable && (
@@ -194,7 +198,7 @@ export const InformationPanel: FC<Props> = ({
 
 const MemoizedHeading = memo<
   Pick<Props, 'type'> & {
-    tag: Props['titleTag']
+    unrecommendedTag: Props['unrecommendedTitleTag']
     id: string
     className: string
     children: Props['title']