+++ packages/smarthr-ui/src/components/Text/Text.tsx
@@ -8,6 +8,8 @@ import {
 } from 'react'
 import { type VariantProps, tv } from 'tailwind-variants'
 
+import { useObjectAttributes } from '../../hooks/useObjectAttributes'
+
 import type { AbstractSize, CharRelativeSize } from '../../themes/createSpacing'
 import type { Gap } from '../../types'
 
@@ -140,6 +142,18 @@ const wrapperClassNameGenerator = tv({
   },
 })
 
+type ObjectLabelType = {
+  /** テキスト左に設置するアイコン */
+  prefix?: ReactNode
+  /** テキスト右に設置するアイコン */
+  suffix?: ReactNode
+  /** アイコンと並べるテキストとの溝 */
+  gap?: CharRelativeSize | AbstractSize
+}
+
+type ActualIconType = undefined | ObjectLabelType
+type IconType = ActualIconType | ReactNode
+
 // VariantProps を使うとコメントが書けない〜🥹
 export type TextProps<T extends ElementType = 'span'> = VariantProps<typeof classNameGenerator> & {
   /** テキストコンポーネントの HTML タグ名。初期値は span */
@@ -148,20 +162,16 @@ export type TextProps<T extends ElementType = 'span'> = VariantProps<typeof clas
   emphasis?: boolean
   /** 見た目の種類 */
   styleType?: StyleType
-  /** テキスト左に設置するアイコン */
-  prefixIcon?: ReactNode
-  /** テキスト右に設置するアイコン */
-  suffixIcon?: ReactNode
-  /** アイコンと並べるテキストとの溝 */
-  iconGap?: CharRelativeSize | AbstractSize
+  /** 設置するアイコン */
+  icon?: IconType
 }
+
+const iconObjectConverter = (icon: ReactNode) => (icon ? { prefix: icon } : undefined)
 
 const ActualText = <T extends ElementType = 'span'>({
   emphasis,
   styleType,
-  prefixIcon,
-  suffixIcon,
-  iconGap,
+  icon: orgIcon,
   weight = emphasis ? 'bold' : undefined,
   as: Component = emphasis ? 'em' : 'span',
   size,
@@ -171,8 +181,9 @@ const ActualText = <T extends ElementType = 'span'>({
   whiteSpace,
   className,
   children,
-  ...props
+  ...rest
 }: PropsWithChildren<TextProps<T> & ComponentProps<T>>) => {
+  const icon = useObjectAttributes<IconType, ActualIconType>(orgIcon, iconObjectConverter)
   const actualClassName = useMemo(() => {
     const styleTypeValues = styleType
       ? STYLE_TYPE_MAP[styleType as StyleType]
@@ -189,17 +200,17 @@ const ActualText = <T extends ElementType = 'span'>({
     })
   }, [size, weight, italic, color, leading, whiteSpace, className, styleType])
   const wrapperClassName = useMemo(
-    () => (prefixIcon || suffixIcon ? wrapperClassNameGenerator({ gap: iconGap || 0.25 }) : ''),
-    [prefixIcon, suffixIcon, iconGap],
+    () => (icon ? wrapperClassNameGenerator({ gap: icon.gap || 0.25 }) : ''),
+    [icon],
   )
 
   return (
-    <Component {...props} className={actualClassName}>
-      {prefixIcon || suffixIcon ? (
+    <Component {...rest} className={actualClassName}>
+      {icon ? (
         <span className={wrapperClassName}>
-          {prefixIcon}
+          {icon.prefix}
           {children}
-          {suffixIcon}
+          {icon.suffix}
         </span>
       ) : (
         children