+++ packages/smarthr-ui/CHANGELOG.md
@@ -2,6 +2,105 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [84.0.0](https://github.com/kufu/smarthr-ui/compare/v83.0.1...v84.0.0) (2026-02-05)
+
+
+### ⚠ BREAKING CHANGES
+
+* 基本的に使ってほしくはないが利用することは可能という属性は、unrecommendedXxxという形式にリネーム (#6057)
+* **PageHeading:** Next.js環境ではページの自動設定を無効にするよう修正 (#6047)
+
+### Features
+
+* FaBookmarkIconを追加 ([#6071](https://github.com/kufu/smarthr-ui/issues/6071)) ([6bba6fa](https://github.com/kufu/smarthr-ui/commit/6bba6fa1ad1b6300341fcd89f0827fbe097dccae))
+
+
+### Bug Fixes
+
+* **Calendar:** Dropdown 内に DatePicker を設置した場合に、年を切り替えることが出来ない問題を修正する ([#6056](https://github.com/kufu/smarthr-ui/issues/6056)) ([65f2c7e](https://github.com/kufu/smarthr-ui/commit/65f2c7e9b151a7d5faba40ad4a5ce572a814cfea))
+* DropdownMenuButtonの内部処理でclassNameを受け取れない拡張が施されたButtonを設置するとstyle崩れが起きる問題を修正する ([#6038](https://github.com/kufu/smarthr-ui/issues/6038)) ([a1f6395](https://github.com/kufu/smarthr-ui/commit/a1f63953bc058bf7581f0b9fe77889fb53b956c9))
+* **PageHeading:** Next.js環境ではページの自動設定を無効にするよう修正 ([#6047](https://github.com/kufu/smarthr-ui/issues/6047)) ([41f67d3](https://github.com/kufu/smarthr-ui/commit/41f67d3e324b53791ed1d0476bc47ab0ce7b2bda))
+* ReactIntlProvider が アプリケーション側の onError などを引き継げるようにした ([#6014](https://github.com/kufu/smarthr-ui/issues/6014)) ([8ebb699](https://github.com/kufu/smarthr-ui/commit/8ebb699528557fdf56376a7b74fa10c87536c34d))
+
+
+* 基本的に使ってほしくはないが利用することは可能という属性は、unrecommendedXxxという形式にリネーム ([#6057](https://github.com/kufu/smarthr-ui/issues/6057)) ([2823d11](https://github.com/kufu/smarthr-ui/commit/2823d117a88acfa2a7f811b9d2a1fe1c30157aaf))
+
+### [83.0.1](https://github.com/kufu/smarthr-ui/compare/v83.0.0...v83.0.1) (2026-01-22)
+
+
+### Bug Fixes
+
+* **SideNavItemButton:** classNameを指定しても設定されないバグを修正する ([#6049](https://github.com/kufu/smarthr-ui/issues/6049)) ([dc31c01](https://github.com/kufu/smarthr-ui/commit/dc31c0171a5106bdf7e50ea81f34449e82d9569d))
+* **Table:** Revert "fixedHeadのPropsがある場合、キーボード操作の逆タブ操作でフォーカス中のエレメントが被ってしまう ([#6051](https://github.com/kufu/smarthr-ui/issues/6051)) ([3b4395c](https://github.com/kufu/smarthr-ui/commit/3b4395cb8df69e7a982c2bbc726bdc673b91afb0))
+
+## [83.0.0](https://github.com/kufu/smarthr-ui/compare/v82.1.0...v83.0.0) (2026-01-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* **DropdownMenuButton:** label, trigger関連の属性をtrigger属性一つにまとめる (#6008)
+
+### Features
+
+* DatePickerを内部的に利用する用になるため、onChangeDateを非推奨にし、onChangeを追加する ([#5936](https://github.com/kufu/smarthr-ui/issues/5936)) ([ad34993](https://github.com/kufu/smarthr-ui/commit/ad3499362bef42b951bedbb265b5658ac6f19214))
+* **DropdownMenuButton:** label, trigger関連の属性をtrigger属性一つにまとめる ([#6008](https://github.com/kufu/smarthr-ui/issues/6008)) ([9d91934](https://github.com/kufu/smarthr-ui/commit/9d919340ae717da8c3aee79176f93bdd88a36690))
+* **SideNavItemButton:** suffix属性を実装 ([#6043](https://github.com/kufu/smarthr-ui/issues/6043)) ([6a6db3c](https://github.com/kufu/smarthr-ui/commit/6a6db3cccd63df4929b5e11df3243e1511b44763))
+
+
+### Bug Fixes
+
+* AppHeaderの言語切替ボタンでの矢印キーの動作を修正 ([#6041](https://github.com/kufu/smarthr-ui/issues/6041)) ([f623792](https://github.com/kufu/smarthr-ui/commit/f6237921fa5162802b028fcb425be0596253fdac))
+* AppHeader内のAppNaviCustomTagのスタイル修正 ([#6042](https://github.com/kufu/smarthr-ui/issues/6042)) ([298ae9b](https://github.com/kufu/smarthr-ui/commit/298ae9b8129c8eae3e362c6205b0c44b3b86ba9e))
+
+## [82.1.0](https://github.com/kufu/smarthr-ui/compare/v82.0.0...v82.1.0) (2026-01-08)
+
+
+### Features
+
+* **DateFormatter:** date属性にstring型で日付を指定できるように修正 ([#6019](https://github.com/kufu/smarthr-ui/issues/6019)) ([aee6509](https://github.com/kufu/smarthr-ui/commit/aee65095eeed982254e7ed99b137f7ba9609f7a1))
+* **FormControl:** 内部の入力要素に対してidを指定した場合でもlabelと正しく紐づくように修正 ([#6034](https://github.com/kufu/smarthr-ui/issues/6034)) ([d0a45d2](https://github.com/kufu/smarthr-ui/commit/d0a45d2c6d66d3f708fff81fa3049ffd0c58533c))
+* **UpwardLink:** モバイル対応 ([#6032](https://github.com/kufu/smarthr-ui/issues/6032)) ([3db7656](https://github.com/kufu/smarthr-ui/commit/3db76560b228a2b98c4bd34cdff2a2484ec1bf8f))
+* useEnvironment/EnvironmentProviderとmediaQueryトークンを追加 ([#5954](https://github.com/kufu/smarthr-ui/issues/5954)) ([65761dd](https://github.com/kufu/smarthr-ui/commit/65761dd5cabd0a340be045b2db10650e29fba93a))
+
+## [82.0.0](https://github.com/kufu/smarthr-ui/compare/v81.0.0...v82.0.0) (2025-12-18)
+
+
+### ⚠ BREAKING CHANGES
+
+* **InputFile:** multipleAppendableを multiple={{ appendable: true }}という形式で指定するように修正 (#5992)
+
+### Features
+
+* **InputFile:** multipleAppendableを multiple={{ appendable: true }}という形式で指定するように修正 ([#5992](https://github.com/kufu/smarthr-ui/issues/5992)) ([917267d](https://github.com/kufu/smarthr-ui/commit/917267d3f30310165d2c99a132e146ae23dc3002))
+
+
+### Bug Fixes
+
+* **Table:** fixedHeadのPropsがある場合、キーボード操作の逆タブ操作でフォーカス中のエレメントが被ってしまう ([#6009](https://github.com/kufu/smarthr-ui/issues/6009)) ([02e7f4b](https://github.com/kufu/smarthr-ui/commit/02e7f4bf4f74b063a28c9a28559c8208cf843341))
+
+## [81.0.0](https://github.com/kufu/smarthr-ui/compare/v80.0.0...v81.0.0) (2025-12-12)
+
+
+### ⚠ BREAKING CHANGES
+
+* **Icon:** FaUpRightFromSquareIconを削除し、OpenInNewTabIconとしてaltのデフォルト値が設定されている状態でexportする (#5991)
+* remove ErrorScreen (#5978)
+* **RadioButtonPanel:** labelSuffix属性を削除し、Clusterを利用してもらう方式に変更 (#5982)
+* **Icon:** 非推奨など利用してほしくないアイコンを削除する (#5972)
+
+### Features
+
+* FaLocationDot, FaMugHot, FaPersonWalkingArrowLoopLeft アイコンを追加 ([#5973](https://github.com/kufu/smarthr-ui/issues/5973)) ([4a2c219](https://github.com/kufu/smarthr-ui/commit/4a2c219ad86393687fc4ecfab25b240c22a79cc5))
+* **Icon:** FaUpRightFromSquareIconを削除し、OpenInNewTabIconとしてaltのデフォルト値が設定されている状態でexportする ([#5991](https://github.com/kufu/smarthr-ui/issues/5991)) ([5e51885](https://github.com/kufu/smarthr-ui/commit/5e51885b79a63d53529fb6231d8ea1bc74f7e2c8))
+* **RadioButtonPanel:** labelSuffix属性を削除し、Clusterを利用してもらう方式に変更 ([#5982](https://github.com/kufu/smarthr-ui/issues/5982)) ([0140e67](https://github.com/kufu/smarthr-ui/commit/0140e6703afcee95e8fbf019f9e65d0b060bc832))
+* remove ErrorScreen ([#5978](https://github.com/kufu/smarthr-ui/issues/5978)) ([b0a39ce](https://github.com/kufu/smarthr-ui/commit/b0a39ce46bba98a1642b20ccaeaf20704a5805d4))
+
+
+### Bug Fixes
+
+* **BulkActionRow:** colSpanを初期表示時999にすることで横幅がTableと揃うようにする ([#5974](https://github.com/kufu/smarthr-ui/issues/5974)) ([9ca8ec0](https://github.com/kufu/smarthr-ui/commit/9ca8ec00df4d7e03578ff0eae4b17db5c2c26251))
+* **Icon:** 非推奨など利用してほしくないアイコンを削除する ([#5972](https://github.com/kufu/smarthr-ui/issues/5972)) ([448c21c](https://github.com/kufu/smarthr-ui/commit/448c21c007723621a8eb880431289b33e8b83647))
+
 ## [80.0.0](https://github.com/kufu/smarthr-ui/compare/v79.2.0...v80.0.0) (2025-12-04)